adds Coin enum:

+++ src/models/Rectangle.java
@@ -3,8 +3,8 @@ package models;
 import models.interfaces.Polygon;
 
 public class Rectangle implements Polygon {
-    private final int length;
-    private final int width;
+    private final double length;
+    private final double width;
 
     public Rectangle(double length, double width) {
         this.length = length;
